+++ craft_parts/plugins/npm_plugin.py
@@ -283,6 +283,7 @@ class NpmPlugin(Plugin):
         # set the Node environment to production mode
         base_env = {
             "NODE_ENV": "production",
+            "NODE_USE_SYSTEM_CA": "1",
         }
         if cast(NpmPluginProperties, self._options).npm_include_node:
             base_env["PATH"] = "${CRAFT_PART_INSTALL}/bin:${PATH}"